Upgrade code snippet to latest version
As part of this upgrade, there is an error check that ensures there are
no namepsacing conflicts already attached to window.

As a result, I needed to reorganize the specs for the client to ensure
that when we validate the script correctly attaches itself to window,
they don't fail as a result of the mock namespace already being present.